requirements.txt
polars>=0.20.0
google-cloud-bigquery>=3.0.0
google-cloud-storage>=3.0.0
matplotlib>=3.0.0
seaborn>=0.12.0
pandas>=1.0.0
db-dtypes>=1.0.0
scikit-learn>=1.3.0
gcsfs>=2024.1.0
faiss-cpu>=1.7.4
